pyproject.toml
[tool.poetry]
name = "anisearch"
version = "0.1.0"
description = ""
authors = ["ssslakter ", "cortuzz "]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
click = "^8.1.3"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"
torch = { version = "2.0.1", source = "torch-cuda-11-8" }
torchvision = { version = "0.15.2", source = "torch-cuda-11-8" }
langchain = { extras = ["qdrant"], version = "0.0.186" }
langchain-serve = "^0.0.43"
dataclass-wizard = "^0.22.2"
fastapi = "^0.96.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.3"
ipykernel = "^6.22.0"
pytest = "^7.3.1"
ipywidgets = "^8.0.6"
wandb = "^0.15.4"

[[tool.poetry.source]]
name = "torch-cuda-11-8"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-line-length = 120
disable = [
    'import-error',
    'too-few-public-methods',
    'missing-module-docstring',
    'too-many-locals',
    'invalid-name',
    'too-many-instance-attributes',
]
extension-pkg-whitelist = 'pydantic'
fail-under = 9.5
ignore-patterns = 'test_.*\.py'
